Use common code for inheritance between parent and child fields
in the createNewFieldEditor function and in the createGroup method.

The common code is in the `inheritField` function which specifies a
general behavior of inheritance between parent fields (which are
currently only of type "object", maybe later we will also implement the
inheritance mechanism for the type "array") and child fields (of
elementary types and maybe also of types "object" and in the future
"array"). Currently the `inheritField` function is used also when the
input groups are created from the user interface, and of course also in
the code that creates the input groups from the specified schema object.