Implemented adding columns of type "object" to array tables from the UI
in all of the three cases (0, 1 or more already existing columns) but
there are many bugs and a lot of work to be done. Also made the creation
of fields of elementary types inside objects under a column of type
"object" to be reflected in all the objects under that column. Also
created a new private function: `fieldPathIsTableColumn`.